Boosting speed of pathfinding.  2d pathfinding fix included not using chebyshev when no diagonals are allowed, manhattan instead.
