+++ src/main/java/com/stewsters/util/pathing/twoDimention/shared/PathNode2d.java
@@ -44,6 +44,10 @@ public class PathNode2d implements Comparable<PathNode2d> {
             return -1;
         } else if (f > of) {
             return 1;
+        } else if (cost > other.cost) {
+            return -1;
+        } else if (cost < other.cost) {
+            return 1;
         } else {
             return 0;
         }
